titanic.py INFO
The model tries to guess if a passenger will survive or not to Titanic, given a big dataset. For this purpose, we are taking only the information related to the age, sex and ticket-class of each individual, and fix some of the missing values on each column to avoid problems with the model. The final model, after almost 280 iterations, gets an accuracy of 0.76315